+++ app/src/main/java/bs/joker/weatherforecast/model/weatherbitio/hourly/DatumHourlyWeatherbitio.java
@@ -43,7 +43,7 @@ public class DatumHourlyWeatherbitio extends RealmObject {
     private String timestampLocal;
     @SerializedName("snow_depth")
     @Expose
-    private Integer snowDepth;
+    private Double snowDepth;
     @SerializedName("clouds")
     @Expose
     private Integer clouds;
@@ -70,7 +70,7 @@ public class DatumHourlyWeatherbitio extends RealmObject {
     private Double dewpt;
     @SerializedName("snow")
     @Expose
-    private Integer snow;
+    private Double snow;
     @SerializedName("uv")
     @Expose
     private Double uv;
@@ -196,11 +196,11 @@ public class DatumHourlyWeatherbitio extends RealmObject {
         this.timestampLocal = timestampLocal;
     }
 
-    public Integer getSnowDepth() {
+    public Double getSnowDepth() {
         return snowDepth;
     }
 
-    public void setSnowDepth(Integer snowDepth) {
+    public void setSnowDepth(Double snowDepth) {
         this.snowDepth = snowDepth;
     }
 
@@ -268,11 +268,11 @@ public class DatumHourlyWeatherbitio extends RealmObject {
         this.dewpt = dewpt;
     }
 
-    public Integer getSnow() {
+    public Double getSnow() {
         return snow;
     }
 
-    public void setSnow(Integer snow) {
+    public void setSnow(Double snow) {
         this.snow = snow;
     }
